init stage vereinfacht

+++ src/teensy.h
@@ -1,7 +1,7 @@
 #ifndef TEENSY_H
 #define TEENSY_H
 
-uint8_t teensy_register_dataref(uint8_t ident, char *str, uint8_t type, void (*cb)(uint8_t, uint32_t));
+int32_t teensy_register_dataref(uint8_t ident, char *str, uint8_t type, void (*cb)(uint8_t, uint32_t));
 void teensy_send_int(uint16_t ident, uint32_t d);
 void teensy_send_float(uint16_t ident, float d);
 void teensy_send_command_once(uint16_t id);
